+++ linux/final_version/include/scene.h
@@ -6,7 +6,7 @@
 /*   By: mraasvel <mraasvel@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/12/26 11:39:49 by mraasvel      #+#    #+#                 */
-/*   Updated: 2021/02/03 17:39:10 by mraasvel      ########   odam.nl         */
+/*   Updated: 2021/02/05 00:18:02 by mraasvel      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ typedef struct	s_camera
 {
 	t_vec3		pos;
 	t_vec3		dir;
-	float		fov;
+	double		fov;
 	t_cspace	cspace;
 }				t_camera;
