+++ linux/final_version/include/vectors.h
@@ -6,7 +6,7 @@
 /*   By: mraasvel <mraasvel@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/12/26 11:39:49 by mraasvel      #+#    #+#                 */
-/*   Updated: 2021/02/04 20:27:36 by mraasvel      ########   odam.nl         */
+/*   Updated: 2021/02/05 00:18:19 by mraasvel      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@
 
 typedef struct	s_vec3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }				t_vec3;
 
 typedef struct	s_cspace
@@ -38,16 +38,16 @@ typedef struct	s_m33
 }				t_m33;
 
 t_vec3			vec_add(t_vec3 a, t_vec3 b);
-t_vec3			vec_scalar(float scalar, t_vec3 v);
+t_vec3			vec_scalar(double scalar, t_vec3 v);
 t_vec3			vec_cross(t_vec3 a, t_vec3 b);
-float			vec_dot(t_vec3 a, t_vec3 b);
-t_vec3			vec_gen(float x, float y, float z);
+double			vec_dot(t_vec3 a, t_vec3 b);
+t_vec3			vec_gen(double x, double y, double z);
 void			vec_normalize(t_vec3 *v);
 t_vec3			vec_normalized(t_vec3 v);
-float			vec_len(t_vec3 v);
+double			vec_len(t_vec3 v);
 t_vec3			vec_sub(t_vec3 a, t_vec3 b);
 t_bool			vec_parallel(t_vec3 a, t_vec3 b);
-float			vec_sqrd(t_vec3 v);
+double			vec_sqrd(t_vec3 v);
 void			vec_invert(t_vec3 *v);
 t_vec3			vec_inverted(t_vec3 v);
 
@@ -57,8 +57,8 @@ t_m33			matrix_transpose(t_m33 matrix);
 t_vec3			matrix_vec_mult(t_m33 m, t_vec3 v);
 t_m33			matrix_gen(t_vec3 c1, t_vec3 c2, t_vec3 c3);
 
-float			rad_to_deg(float radians);
-float			deg_to_rad(float degrees);
-float			quadratic_equation(float a, float b, float c);
+double			rad_to_deg(double radians);
+double			deg_to_rad(double degrees);
+double			quadratic_equation(double a, double b, double c);
 
 #endif
